Added calibration correction code for accel and gyro to Windrose module.
Added calibration correction code for accel into the UpdatePosition
function. Added calibration correction code for gyro int the
UpdateHeading function. Still need to add correction code for the mag,
but I am waiting for the PCB interconnect to be done in order to get the
mag data for calculating the calibration constants.